+++ todoist_api_python/api.py
@@ -1333,7 +1333,7 @@ class TodoistAPI:
         name: Annotated[str, MinLen(1), MaxLen(60)],
         *,
         color: ColorString | None = None,
-        item_order: int | None = None,
+        order: int | None = None,
         is_favorite: bool | None = None,
     ) -> Label:
         """
@@ -1341,7 +1341,7 @@ class TodoistAPI:
 
         :param name: The name of the label.
         :param color: The color of the label icon.
-        :param item_order: Label's order in the label list.
+        :param order: Label's order in the label list.
         :param is_favorite: Whether the label is a favorite.
         :return: The newly created label.
         :raises httpx.HTTPStatusError: If the API request fails.
@@ -1352,7 +1352,7 @@ class TodoistAPI:
         data = kwargs_without_none(
             name=name,
             color=color,
-            item_order=item_order,
+            order=order,
             is_favorite=is_favorite,
         )
 
@@ -1372,7 +1372,7 @@ class TodoistAPI:
         *,
         name: Annotated[str, MinLen(1), MaxLen(60)] | None = None,
         color: ColorString | None = None,
-        item_order: int | None = None,
+        order: int | None = None,
         is_favorite: bool | None = None,
     ) -> Label:
         """
@@ -1383,7 +1383,7 @@ class TodoistAPI:
         :param label_id: The ID of the label.
         :param name: The name of the label.
         :param color: The color of the label icon.
-        :param item_order: Label's order in the label list.
+        :param order: Label's order in the label list.
         :param is_favorite: Whether the label is a favorite.
         :return: the updated Label.
         :raises httpx.HTTPStatusError: If the API request fails.
@@ -1393,7 +1393,7 @@ class TodoistAPI:
         data = kwargs_without_none(
             name=name,
             color=color,
-            item_order=item_order,
+            order=order,
             is_favorite=is_favorite,
         )